IC-7-6 Category, Transacion and Settings Repositories

+++ app/src/main/java/com/xinosluitsnoi/mymoney/domain/repository/CategoryRepository.java
@@ -10,7 +10,7 @@ import androidx.annotation.Nullable;
 public interface CategoryRepository {
 
     @Nullable
-    Category getCategory(int id);
+    Category getCategory(long id);
 
     @NonNull
     List<Category> getAll(@Category.Type.Mode int type);
